Move the usercluster controller
