+++ api/pkg/controller/user-cluster-controller-manager/resources/reconciler.go
@@ -5,20 +5,20 @@ import (
 	"fmt"
 
 	openshiftresources "github.com/kubermatic/kubermatic/api/pkg/controller/openshift/resources"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/clusterautoscaler"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/controller-manager"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/dnat-controller"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/kube-state-metrics"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/kubernetes-dashboard"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/machine-controller"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/metrics-server"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/openshift"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/openvpn"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/prometheus"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/scheduler"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/system-basic-user"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/user-auth"
-	"github.com/kubermatic/kubermatic/api/pkg/controller/usercluster/resources/usersshkeys"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/clusterautoscaler"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/controller-manager"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/dnat-controller"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/kube-state-metrics"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/kubernetes-dashboard"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/machine-controller"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/metrics-server"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/openshift"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/openvpn"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/prometheus"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/scheduler"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/system-basic-user"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/user-auth"
+	"github.com/kubermatic/kubermatic/api/pkg/controller/user-cluster-controller-manager/resources/resources/usersshkeys"
 	"github.com/kubermatic/kubermatic/api/pkg/resources/certificates/triple"
 	"github.com/kubermatic/kubermatic/api/pkg/resources/reconciling"
