+++ api/pkg/controller/user-cluster-controller-manager/resources/resources/kube-state-metrics/clusterrole.go
