+++ api/pkg/controller/user-cluster-controller-manager/resources/resources/kubernetes-dashboard/deployment.go
