+++ api/pkg/controller/user-cluster-controller-manager/resources/resources/kubernetes-dashboard/role.go
