+++ api/pkg/controller/user-cluster-controller-manager/resources/resources/machine-controller/cluster-info-configmap.go
