+++ api/pkg/controller/user-cluster-controller-manager/resources/resources/openshift/namespace.go
